I've successfully made the editor behaviour, which changes classes.
What's left to do, is to load all seats' information with ids, rows, cols and type into an array and send it to the API.

After that, I'll only have to copy the code into seat selector for public users to choose seats, create user payment form and create an api route for downloading the ticket. That's it.